#7309: move the new test out to keep the file under the length cap
